+++ sites/example-project/src/components/ui/stores.js
@@ -2,3 +2,4 @@ import { dev } from '$app/env';
 import { writable } from 'svelte/store';
 
 export const showQueries = writable(dev);
+export const pageHasQueries = writable();