+++ packages/db-orchestrator/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @evidence-dev/db-orchestrator
 
+## 1.0.0-next.4
+
+### Patch Changes
+
+- Updated dependencies [3d0d93c]
+  - @evidence-dev/sqlite@1.0.0-next.4
+
 ## 1.0.0-next.3
 
 ### Patch Changes