+++ packages/preprocess/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @evidence-dev/preprocess
 
+## 1.0.0-next.1
+
+### Patch Changes
+
+- 3d0d93c: Various UI changes
+
 ## 1.0.0-next.0
 
 ### Major Changes