+++ sites/test-env/CHANGELOG.md
@@ -1,5 +1,26 @@
 # evidence-test-environment
 
+## 1.0.0-next.8
+
+### Patch Changes
+
+- Updated dependencies [e45a29a]
+  - @evidence-dev/evidence@1.0.0-next.8
+
+## 1.0.0-next.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @evidence-dev/evidence@1.0.0-next.7
+
+## 1.0.0-next.6
+
+### Patch Changes
+
+- Updated dependencies [3d0d93c]
+  - @evidence-dev/evidence@1.0.0-next.6
+
 ## 1.0.0-next.5
 
 ### Patch Changes